refactor: move RULE_ID to rule-901.ts
